Questionable webcam aspect ratio fix
Al we really want is to say "give me the webcam, at its native size", but specifying a dimension, mandatory or ideal, causes browsers to resize and crop the image. This seems to be an unstable standard and is implemented differently between browsers. Specifying just one dimension, of width but not height, seems to avoid cropping the image in the tests I've been able to do. No doubt this will break in the future.